player: packer: Optimise default volume of instruments
When an instrument is specified with no Cxx effect, the default
instrument volume needs to be used. Instead of letting the player
detect this case, do it in the packer, and add a volume entry with the
default volume.

The entries in the patterns have been reordered to simplify the logic.