player: Fix multiple effects in one row not working
In some cases, when multiple channels have effects in the same row, the
last ones would be ignored. This was happening because of some returns
in the modified function by this patch, which should have been continue.
It didn't happen with all effects, which made it harder to detect.

This bug was introduced in commit 0654fff14f44 ("packer: player:
Implement MOD Porta + Vol Slide (5xy)").