player: Fix effect Pattern Break (Dxy)
If the effect was done in any channel but the last one, the pattern
read pointer would be updated before the last channels in that row had
time to process their notes/effects. This commit fixes it by waiting
until all channels have been handled to execute the Pattern Break
effect.

Also, this means that if two channels have a Dxy effect, the last one is
the one that will be executed.