packer: player: tests: Implement MOD E4x and E7x
Implement effects Vibrato Waveform and Tremolo Waveform. The two effects
are the same, except the former one affects frequency and the later one
affects amplitude.

The results of the tests have been verified with Audacity in the case
of the tremolo test. The code is the same in the vibrato side, so it
should also be fine.