player: tests: Reverse period logic in mixer
Instead of keeping track of how many samples it's going to take to
advance the waveform pointer, keep the speed as a fixed point value
that is added to the current position, which is also fixed point.

This removes one conditional per iteration in the mixing loop.

This commit breaks the tests, they will be fixed in another commit.